add stars and fix camera & starship motion manually
starfield of Claude's code is invisible(lol) because stars line length are all ZERO!
and it's some kind of warp effect, I decided to make star a unit line segment to express stream of light.

initial movement is just circular around starship. change it to follow the flight line of starship.